backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# API Clients
tweepy==4.14.0
python-telegram-bot==20.6
praw==7.7.1
web3==6.11.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
textblob==0.17.1

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# Cache & Message Queue
redis==5.0.1
aio-pika==9.3.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0 